Move exclude_keys() to dlt.common.utils module
