+++ dlt/common/utils.py
@@ -504,6 +504,19 @@ def without_none(d: Mapping[TKey, Optional[TValue]]) -> Mapping[TKey, TValue]:
     return {k: v for k, v in d.items() if v is not None}
 
 
+def exclude_keys(d: Mapping[str, Any], keys: Iterable[str]) -> Dict[str, Any]:
+    """Removes specified keys from a dictionary and returns a new dictionary.
+
+    Args:
+        d (Mapping[str, Any]): The dictionary to remove keys from.
+        keys (Iterable[str]): The keys to remove.
+
+    Returns:
+        Dict[str, Any]: A new dictionary with the specified keys removed.
+    """
+    return {k: v for k, v in d.items() if k not in keys}
+
+
 def get_full_class_name(obj: Any) -> str:
     cls = obj.__class__
     module = cls.__module__
